Split mmu.h files into mmu.h and mmu_defs.h.
Keep mmu_defs.h clean, so as to allow it to be included in arbitrary
code.  mmu_defs are to simply provide definitions of things such as
PGSIZE.  mmu.h can contain function definitions and code.

(Without this only rhype code can use mmu.h, as there are function
definitions which conflict in other contexts --- definition of sync)