Add a counters facility that lets one detect how many times a point in the code has been hit.
Up to 16 counter can be active at any time.  When counters are
activated a "li" instruction is patched up to indicate which counter
this particular spot in the code should increment.

pre_link.c and post_link.c have been added to be able to define
symbols in specific sections, thus ensuring that we can get the bounds
of an ELF-section we define. This is needed to allow us to use an elf
section to accumulate the locations of counters throughout the code;
thus letting us know where the li instructions that need patching are.