Various clean-up/refactoring.
Add support for PTE aging in the HPTE.  We tag each HPTE with the
low-order two bits of the current "generation" at the time of
insertion (using 2 of the software bits).  A scanning procedure
applied periodically to the htab looks at these bits and compares them
to the current generation.  If they are out of date wrt the current
generation, the PTE is invalidated.  When the entire htab has been
scanned, the generation count is updated.

Thus if we have inserted a PTE when the generation count is x, and
don't touch if after, we know that the PTE has been evicted if the
generation count is x + 4.

We tag each vm_class with the generation at the time of the last
insertion, thus allowing us to know when a given generation has been
entirely removed.

The mechanism is activated by defining HPTE_AGING (in one of the
htab.h files).  It is currently de-activated --- one could make it
work, but without some extra work in the generic Linux code, Linux
could use this too aggressively (i.e. dropping and aging an entire
vm_class when unmapping on a fork, which is probably overkill).


Implementing this mechanism has required the imposition of some
discipline in using the SW bits in an HPTE:

locked -> don't touch
bolted -> ignore for the purposes of HPTE aging

The remaining two bits are used for the 2-bit generation id.

The vregs mapping permanently locks the HPTE which it flips the
protection bits on (to simulate MER).

All other code uses pte_trylock, instead of pte_lock when locking a
valid HPTE.  (We use pte_lock when going after empty HPTE's or when
we're purging a vm_class out of the htab, in which case we know we
can't be hitting a permanently locked HPTE.)

Using trylock turns out to be a good idea anyway, since in most cases
the fact that an HPTE is locked is a hit to tell us that it is
maybe a bad idea to go after it anyways.