Use a struct mem_ranges to track memory given to partitions.
In fact, use two of them; one for the memory given to partitions as
seen by the controller code, and one describing the memory as seen by
the partition. (The two partition's view of the memory is via
different sets of logical addresses, and the OF tree construction must
use the view as seen by the partition being created.)

Tested for reloader only.  Assuming non-reloader controller will not
be taking advantage of this, and in the case that only a single chunk
is being given, it shouldn't make a difference anyways.

For reloader, it lets us give the controller OS access to more memory
than a single chunk, as the OF tree can now describe more memory.