Define a proper SLB cache object.  The SLB cache object must contain a bitmap identifying valid entries.  We can then use the bitmap to control loops over the cache. (Experimentally I've found that there is a significant performacne impact in doing so.)
Use 2 SLB caches per thread.  One of these caches is used to cache
kernel-mode SLB's.  This allows us to invalidate/re-validate SLB
entries belonging to the kernel as a partition switches between app
and kernel context.  This caching is done independent of the VM
Classes involved, and this too has been observed to provide a big
impact, experimentally (10% on SDET).

Re-write the enter/ext kernel routines to adapt to this scheme.

Make all SLB ops in the code-base use SLB cache objects -- the goal is
to avoid the need to read SLB entires (which is more expensive --
benh).