Re-implement exception vector paths with an alternate implementation.
There are now two implementations:

unified:  does a "ba" at the exception entry point

split:  tries to do a bit of work before jumping

The split method requires using the exc_num field in the
thread_control_area to save the exception number over the jump.  But,
it packs the code in a bit tighter.

For some exceptions we want to use the "split" method explicitly (e.g. machine
check and alignment, which all jump to a common handler)

The default is selected depending on whether or not "SPLIT_PROLOG" is defined.

Eventually one of these will go away.